Every tests passes. Need to go to 101 to validate

+++ src/str/ft_llitoa_base.c
@@ -10,10 +10,10 @@ char	ftllibc(int i)
 
 char	*ft_llitoa_base(long long int nbr, long long int base)
 {
-	int		tmp;
-	int		i;
-	int		mod;
-	char	*str;
+	long long int	tmp;
+	int				i;
+	int				mod;
+	char			*str;
 
 	tmp = nbr;
 	i = 1;
